Move UX family to common folder
